fix: incorrect path arrangement (#7)
