+++ work1/RukiaOvO/AtCoder_ARC017A.go
