+++ work1/RukiaOvO/Bonus3[map].go
